Visual Studio build maintenance.
* Fix compiler warnings
* Add Visual Studio 2013 solution
* Tweak .pro files
* Most importantly, rename main 'quackle' project to
  'libquackle' to avoid solution-level name conflict.  This
  only affects the VS build.

+++ datamanager.cpp
@@ -42,7 +42,7 @@ DataManager::DataManager()
 {
 	m_self = this;
 	setDataDirectory(".");
-    seedRandomNumbers(time(NULL));
+    seedRandomNumbers((int)time(NULL));
 
 	m_alphabetParameters = new EnglishAlphabetParameters;
 	m_evaluator = new CatchallEvaluator;
